Calculate gene flow (Fsp)

Calculate gene flow (Fsp)

get_facility_fsp(fasta, locs, matrix = TRUE)

Arguments

fasta
ape DNAbin object (i.e. from fasta file of SNPs) using read.fasta

locs
a named vector of locations of isolates (e.g. facility of isolation), with the name being the sample ID

matrix
whether to output symmetric matrix (TRUE; default) or long form (FALSE)

Value

facility x facility matrix with Fsp values

Details

Fsp is described in Donker et al. 2017 (mgen.microbiologyresearch.org/pubmed/content/journal/mgen/10.1099/mgen.0.000113). Only bi-allelic sites are included when computing Fsp. The Fsp values are between 0 and 1 where lower values indicate more similar populations. Note that the current implementation of this function is fairly slow.

Examples

if (FALSE) {
# This takes a long time to run right now!
locs <- metadata %>% dplyr::select(isolate_id, facility) %>% tibble::deframe()
facil_fsp <- get_facility_fsp(aln, locs, matrix = TRUE)
}