Summarize inter-facility patient transfer network

Summarize inter-facility patient transfer network from an edge list. Direct and indirect patient flow metrics are calculated.

get_patient_flow(pt_trans_df, paths = FALSE)

Arguments

pt_trans_df
a dataframe representing a patient transfer network of 3 cols: 'source_facil', 'dest_facil, and 'n_transfers' (code doesn't support missing paths, any missing paths will be represented by 0s)

paths
boolean value, TRUE if you want the shortest paths returned, FALSE if you don't

Value

the number of direct patient transfers and indirect flow metrics between each facility pair. If paths = TRUE, a list of summary (pt_trans_summary) and shortest paths used (paths).

Details

For more details on how patient flow is calculated, see: https://aac.asm.org/content/63/11/e01622-19.

Examples

get_patient_flow(pt_trans_df = pt_trans_df)

#>     loc1 loc2 n_transfers_f12 pt_trans_metric_f12 n_transfers_f21
#> 1      D    H             260         0.045129566              17
#> 2      D    G               5         0.032123396              14
#> 3      C    D             474         0.062188402            2660
#> 4      D    T               8         0.026630252             182
#> 5      D    F              46         0.066839138              60
#> 6      D    I              12         0.164514217              63
#> 7      D    E             504         0.088158274             330
#> 8      D    P               8         0.022960386              18
#> 9      D    M            3292         0.484046464            4244
#> 10     D    O               6         0.003244662              38
#> 11     D    H             260         0.045129566              17
#> 12     G    H              21         0.002670972              10
#> 13     C    H             492         0.115385781             655
#> 14     H    T            2797         0.590084388            2640
#> 15     F    H              28         0.016010519              33
#> 16     H    I             769         0.258898331            1957
#> 17     E    H              63         0.020248344             124
#> 18     H    P               8         0.036133082              27
#> 19     H    M             319         0.067299578             255
#> 20     H    O               8         0.005106170             118
#> 21     D    G               5         0.032123396              14
#> 22     G    H              21         0.002670972              10
#> 23     C    G              22         0.082132035             189
#> 24     G    T               3         0.001576099              43
#> 25     F    G             935         0.066425121            2183
#> 26     G    I             142         0.009736698            1393
#> 27     E    G              76         0.050361670            1438
#> 28     G    P           10424         0.714755897            5372
#> 29     G    M              45         0.003716879              18
#> 30     G    O             125         0.101006196            2035
#> 31     C    D             474         0.062188402            2660
#> 32     C    H             492         0.115385781             655
#> 33     C    G              22         0.082132035             189
#> 34     C    T              90         0.068087348              65
#> 35     C    F             563         0.170892098             832
#> 36     C    I            3206         0.420624508            2703
#> 37     C    E            1718         0.225400157            1632
#> 38     C    P               8         0.058704357              24
#> 39     C    M            1043         0.136840724           10638
#> 40     C    O               6         0.008295844            1050
#> 41     D    T               8         0.026630252             182
#> 42     H    T            2797         0.590084388            2640
#> 43     G    T               3         0.001576099              43
#> 44     C    T              90         0.068087348              65
#> 45     F    T              20         0.009447557              44
#> 46     I    T             378         0.161872042            2439
#> 47     E    T               3         0.011948231              44
#> 48     P    T               2         0.002071279              26
#> 49     M    T               5         0.045528519              44
#> 50     O    T              52         0.012582676              32
#> 51     D    F              46         0.066839138              60
#> 52     F    H              28         0.016010519              33
#> 53     F    G             935         0.066425121            2183
#> 54     C    F             563         0.170892098             832
#> 55     F    T              20         0.009447557              44
#> 56     F    I             596         0.058364354              95
#> 57     E    F            7051         0.758172043           11130
#> 58     F    P             405         0.047477747             989
#> 59     F    M              10         0.018987530              45
#> 60     F    O              60         0.006709349             102
#> 61     D    I              12         0.164514217              63
#> 62     H    I             769         0.258898331            1957
#> 63     G    I             142         0.009736698            1393
#> 64     C    I            3206         0.420624508            2703
#> 65     I    T             378         0.161872042            2439
#> 66     F    I             596         0.058364354              95
#> 67     E    I              20         0.073812817             191
#> 68     I    P             333         0.139564755              18
#> 69     I    M              10         0.051847558              34
#> 70     I    O              11         0.019722684             569
#> 71     D    E             504         0.088158274             330
#> 72     E    H              63         0.020248344             124
#> 73     E    G              76         0.050361670            1438
#> 74     C    E            1718         0.225400157            1632
#> 75     E    T               3         0.011948231              44
#> 76     E    F            7051         0.758172043           11130
#> 77     E    I              20         0.073812817             191
#> 78     E    P              30         0.035996300             115
#> 79     E    M              77         0.024013340             494
#> 80     E    O              18         0.005086841             235
#> 81     D    P               8         0.022960386              18
#> 82     H    P               8         0.036133082              27
#> 83     G    P           10424         0.714755897            5372
#> 84     C    P               8         0.058704357              24
#> 85     P    T               2         0.002071279              26
#> 86     F    P             405         0.047477747             989
#> 87     I    P             333         0.139564755              18
#> 88     E    P              30         0.035996300             115
#> 89     M    P             172         0.039254318              40
#> 90     O    P            2871         0.392213115            1087
#> 91     D    M            3292         0.484046464            4244
#> 92     H    M             319         0.067299578             255
#> 93     G    M              45         0.003716879              18
#> 94     C    M            1043         0.136840724           10638
#> 95     M    T               5         0.045528519              44
#> 96     F    M              10         0.018987530              45
#> 97     I    M              10         0.051847558              34
#> 98     E    M              77         0.024013340             494
#> 99     M    P             172         0.039254318              40
#> 100    M    O               4         0.005547250             250
#> 101    D    O               6         0.003244662              38
#> 102    H    O               8         0.005106170             118
#> 103    G    O             125         0.101006196            2035
#> 104    C    O               6         0.008295844            1050
#> 105    O    T              52         0.012582676              32
#> 106    F    O              60         0.006709349             102
#> 107    I    O              11         0.019722684             569
#> 108    E    O              18         0.005086841             235
#> 109    O    P            2871         0.392213115            1087
#> 110    M    O               4         0.005547250             250
#>     pt_trans_metric_f21 sum_transfers sum_pt_trans_metric
#> 1           0.019319187           277          0.06444875
#> 2           0.004199755            19          0.03632315
#> 3           0.391118953          3134          0.45330736
#> 4           0.032739701           190          0.05936995
#> 5           0.028057366           106          0.09489650
#> 6           0.023562553            75          0.18807677
#> 7           0.035483871           834          0.12364214
#> 8           0.003607480            26          0.02656787
#> 9           0.266767239          7536          0.75081370
#> 10          0.009110903            44          0.01235556
#> 11          0.019319187           277          0.06444875
#> 12          0.050553038            31          0.05322401
#> 13          0.138185654          1147          0.25357144
#> 14          0.474905559          5437          1.06498995
#> 15          0.023614836            61          0.03962536
#> 16          0.274320157          2726          0.53321849
#> 17          0.031147068           187          0.05139541
#> 18          0.003510140            35          0.03964322
#> 19          0.077155946           574          0.14445552
#> 20          0.021323520           126          0.02642969
#> 21          0.004199755            19          0.03632315
#> 22          0.050553038            31          0.05322401
#> 23          0.020769698           211          0.10290173
#> 24          0.085670862            46          0.08724696
#> 25          0.149684586          3118          0.21610971
#> 26          0.195262125          1535          0.20499882
#> 27          0.118356738          1514          0.16871841
#> 28          0.698387936         15796          1.41314383
#> 29          0.054919894            63          0.05863677
#> 30          0.278005464          2160          0.37901166
#> 31          0.391118953          3134          0.45330736
#> 32          0.138185654          1147          0.25357144
#> 33          0.020769698           211          0.10290173
#> 34          0.166237143           155          0.23432449
#> 35          0.138756428          1395          0.30964853
#> 36          0.378889823          5909          0.79951433
#> 37          0.175483871          3350          0.40088403
#> 38          0.020270688            32          0.07897504
#> 39          0.668678107         11681          0.80551883
#> 40          0.143442623          1056          0.15173847
#> 41          0.032739701           190          0.05936995
#> 42          0.474905559          5437          1.06498995
#> 43          0.085670862            46          0.08724696
#> 44          0.166237143           155          0.23432449
#> 45          0.028408614            64          0.03785617
#> 46          0.438747976          2817          0.60062002
#> 47          0.037469878            47          0.04941811
#> 48          0.061233754            28          0.06330503
#> 49          0.031960944            49          0.07748946
#> 50          0.008653288            84          0.02123596
#> 51          0.028057366           106          0.09489650
#> 52          0.023614836            61          0.03962536
#> 53          0.149684586          3118          0.21610971
#> 54          0.138756428          1395          0.30964853
#> 55          0.028408614            64          0.03785617
#> 56          0.064749277           691          0.12311363
#> 57          0.790707587         18181          1.54887963
#> 58          0.128575143          1394          0.17605289
#> 59          0.114271804            55          0.13325933
#> 60          0.050428857           162          0.05713821
#> 61          0.023562553            75          0.18807677
#> 62          0.274320157          2726          0.53321849
#> 63          0.195262125          1535          0.20499882
#> 64          0.378889823          5909          0.79951433
#> 65          0.438747976          2817          0.60062002
#> 66          0.064749277           691          0.12311363
#> 67          0.085401826           211          0.15921464
#> 68          0.010984782           351          0.15054954
#> 69          0.281262400            44          0.33310996
#> 70          0.077732240           580          0.09745492
#> 71          0.035483871           834          0.12364214
#> 72          0.031147068           187          0.05139541
#> 73          0.118356738          1514          0.16871841
#> 74          0.175483871          3350          0.40088403
#> 75          0.037469878            47          0.04941811
#> 76          0.790707587         18181          1.54887963
#> 77          0.085401826           211          0.15921464
#> 78          0.101665341           145          0.13766164
#> 79          0.150720151           571          0.17473349
#> 80          0.039874480           253          0.04496132
#> 81          0.003607480            26          0.02656787
#> 82          0.003510140            35          0.03964322
#> 83          0.698387936         15796          1.41314383
#> 84          0.020270688            32          0.07897504
#> 85          0.061233754            28          0.06330503
#> 86          0.128575143          1394          0.17605289
#> 87          0.010984782           351          0.15054954
#> 88          0.101665341           145          0.13766164
#> 89          0.005200208           212          0.04445453
#> 90          0.141315653          3958          0.53352877
#> 91          0.266767239          7536          0.75081370
#> 92          0.077155946           574          0.14445552
#> 93          0.054919894            63          0.05863677
#> 94          0.668678107         11681          0.80551883
#> 95          0.031960944            49          0.07748946
#> 96          0.114271804            55          0.13325933
#> 97          0.281262400            44          0.33310996
#> 98          0.150720151           571          0.17473349
#> 99          0.005200208           212          0.04445453
#> 100         0.034153005           254          0.03970026
#> 101         0.009110903            44          0.01235556
#> 102         0.021323520           126          0.02642969
#> 103         0.278005464          2160          0.37901166
#> 104         0.143442623          1056          0.15173847
#> 105         0.008653288            84          0.02123596
#> 106         0.050428857           162          0.05713821
#> 107         0.077732240           580          0.09745492
#> 108         0.039874480           253          0.04496132
#> 109         0.141315653          3958          0.53352877
#> 110         0.034153005           254          0.03970026