Make matrix long-form where each row represents a cell in the matrix

Each row represents a cell in the matrix (rowname, column name, value). If the matrix is symmetric, there will be one row for each symmetric pair (i.e. the number of rows in the long-form data frame will be half the number of elements in the input matrix). If the matrix is asymmetric, all rows will be kept (i.e. the number of rows in the long-form data frame will be equal to the number of elements in the input matrix).

make_long_form(facil_dist, col_names = c("loc1", "loc2", "fsp"))

Arguments

facil_dist
symmetric or asymmetric matrix that you want to convert to long form.

col_names
Column names for the output data frame (3 columns - rownames, colnames, values)

Value

long form data where each row represents a cell in the matrix (rowname, column name, value)

Examples

make_long_form(fsp)
#>     loc1 loc2       fsp
#> 1      A    C 0.7717549
#> 2      A    D 0.7663591
#> 3      A    E 0.8194635
#> 4      A    F 0.7594117
#> 5      A    G 0.7383896
#> 6      A    H 0.8410055
#> 7      A    I 0.8667262
#> 8      A    J 0.9764447
#> 9      A    K 0.9522750
#> 10     A    M 0.7662909
#> 11     A    N 0.8133973
#> 12     A    O 0.8306250
#> 13     A    P 0.7425305
#> 14     A    R 0.8115470
#> 15     A    T 0.8319213
#> 16     A    U 0.8421131
#> 18     C    D 0.3413702
#> 19     C    E 0.3415101
#> 20     C    F 0.2628256
#> 21     C    G 0.5334353
#> 22     C    H 0.2403687
#> 23     C    I 0.3084998
#> 24     C    J 0.9839316
#> 25     C    K 0.8827399
#> 26     C    M 0.1504062
#> 27     C    N 0.5465672
#> 28     C    O 0.5735577
#> 29     C    P 0.5341858
#> 30     C    R 0.6547738
#> 31     C    T 0.1884005
#> 32     C    U 0.6816511
#> 35     D    E 0.2434790
#> 36     D    F 0.2536502
#> 37     D    G 0.4939398
#> 38     D    H 0.4006562
#> 39     D    I 0.5973915
#> 40     D    J 0.9805450
#> 41     D    K 0.8396582
#> 42     D    M 0.3663218
#> 43     D    N 0.4586921
#> 44     D    O 0.5319298
#> 45     D    P 0.4289203
#> 46     D    R 0.6282581
#> 47     D    T 0.3786305
#> 48     D    U 0.7077745
#> 52     E    F 0.2514725
#> 53     E    G 0.6655447
#> 54     E    H 0.3056815
#> 55     E    I 0.4980262
#> 56     E    J 0.9718302
#> 57     E    K 0.9182785
#> 58     E    M 0.3001906
#> 59     E    N 0.4222323
#> 60     E    O 0.5653355
#> 61     E    P 0.5857677
#> 62     E    R 0.6903798
#> 63     E    T 0.4121684
#> 64     E    U 0.7565016
#> 69     F    G 0.5616049
#> 70     F    H 0.3484753
#> 71     F    I 0.4898334
#> 72     F    J 0.9778149
#> 73     F    K 0.9050275
#> 74     F    M 0.2484644
#> 75     F    N 0.4849724
#> 76     F    O 0.5135410
#> 77     F    P 0.4670237
#> 78     F    R 0.6282514
#> 79     F    T 0.3535072
#> 80     F    U 0.6913074
#> 86     G    H 0.5908832
#> 87     G    I 0.7489789
#> 88     G    J 0.9756172
#> 89     G    K 0.7230079
#> 90     G    M 0.6895115
#> 91     G    N 0.6774641
#> 92     G    O 0.4742981
#> 93     G    P 0.1962121
#> 94     G    R 0.7805078
#> 95     G    T 0.4781963
#> 96     G    U 0.7814417
#> 103    H    I 0.2133007
#> 104    H    J 0.9842605
#> 105    H    K 0.8892935
#> 106    H    M 0.2853168
#> 107    H    N 0.4023813
#> 108    H    O 0.4058176
#> 109    H    P 0.5102451
#> 110    H    R 0.6948037
#> 111    H    T 0.2846916
#> 112    H    U 0.6896490
#> 120    I    J 0.9910524
#> 121    I    K 0.9906516
#> 122    I    M 0.3524131
#> 123    I    N 0.4794344
#> 124    I    O 0.4275278
#> 125    I    P 0.6598976
#> 126    I    R 0.7683236
#> 127    I    T 0.3896366
#> 128    I    U 0.7394817
#> 137    J    K 0.9904933
#> 138    J    M 0.9831099
#> 139    J    N 0.9872937
#> 140    J    O 0.9862792
#> 141    J    P 0.9792813
#> 142    J    R 0.9835531
#> 143    J    T 0.9868078
#> 144    J    U 0.9900439
#> 154    K    M 0.9488845
#> 155    K    N 0.9106904
#> 156    K    O 0.8565391
#> 157    K    P 0.7548870
#> 158    K    R 0.7527014
#> 159    K    T 0.8148390
#> 160    K    U 0.8847174
#> 171    M    N 0.5354747
#> 172    M    O 0.6820799
#> 173    M    P 0.6291749
#> 174    M    R 0.6616427
#> 175    M    T 0.2138976
#> 176    M    U 0.7054690
#> 188    N    O 0.3983550
#> 189    N    P 0.5732978
#> 190    N    R 0.7414548
#> 191    N    T 0.6161091
#> 192    N    U 0.7787368
#> 205    O    P 0.3988167
#> 206    O    R 0.7791095
#> 207    O    T 0.5617683
#> 208    O    U 0.7373228
#> 222    P    R 0.7462601
#> 223    P    T 0.4855295
#> 224    P    U 0.7660985
#> 239    R    T 0.6481825
#> 240    R    U 0.3539550
#> 256    T    U 0.6691453